Update Monthly Performance Analysis - Graphing.py
Added if completion date is null and the close out date is not null to supplement  the close out date value as the completion date in case a WO was closed out without being completed. This solution is simpler than adjusted further logic to switch to the Close Out Date Column.